feat: Dummy UI with selectable request/response
