+++ ui/src/Call.ts
@@ -1,0 +1,6 @@
+export interface Call {
+    id: number
+    request: Object
+    response: Object
+}
+
